postfinance/requirements.txt
natsort==5.0.2
pdfminer.six==20160614
ruamel.yaml==0.14.5
six==1.10.0